+++ src/components/molecules/Menu/Menu.tsx
@@ -1,13 +1,10 @@
 import React from 'react';
-import axios from 'axios';
-import { toast } from 'react-toastify';
-import { NavLink, Link, useHistory } from 'react-router-dom';
+import { NavLink, Link } from 'react-router-dom';
 import AppBar from '@material-ui/core/AppBar';
 import { makeStyles, Toolbar } from '@material-ui/core';
 import Grid from '@material-ui/core/Grid';
 import Typo from '../../atoms/Typo/Typo';
-import { useAppDispatch, useUserDispatch } from '../../../utils/hooks/useContext';
-import { makeAPIPath } from '../../../utils/utils';
+import useLogout from '../../../utils/hooks/useLogout';
 
 const useStyles = makeStyles({
   cursor: {
@@ -16,10 +13,8 @@ const useStyles = makeStyles({
 });
 
 const Menu = () => {
-  const history = useHistory();
-  const appDispatch = useAppDispatch();
-  const userDispatch = useUserDispatch();
   const classes = useStyles();
+  const handleLogout = useLogout();
 
   const Choices = ['Game', 'DM', 'Channel', 'Community', 'Profile'];
   const RenderChoices = Choices.map((item) => (
@@ -35,21 +30,6 @@ const Menu = () => {
       </NavLink>
     </Grid>
   ));
-
-  const handleLogout = () => {
-    appDispatch({ type: 'loading' });
-    axios.get(makeAPIPath('/auth/logout'))
-      .finally(() => {
-        appDispatch({ type: 'endLoading' });
-      })
-      .then(() => {
-        userDispatch({ type: 'logout' });
-        history.push('/');
-      })
-      .catch((error) => {
-        toast.error(error.message);
-      });
-  };
 
   const menu = RenderChoices.concat(
     <Grid item key="Logout" onClick={handleLogout}>
