+++ src/components/pages/MFAPage/MFAPage.tsx
@@ -96,7 +96,7 @@ const MFAPage = () => {
           info: {
             id,
             name,
-            avatar,
+            avatar: makeAPIPath(`/${avatar}`),
             enable2FA,
             authenticatorSecret,
             isSecondFactorAuthenticated,
